+++ waspc/cli/src/Wasp/Cli/Command/Start.hs
@@ -9,8 +9,8 @@ import Control.Concurrent.MVar (MVar, newMVar, tryTakeMVar)
 import Control.Monad.Except (throwError)
 import Control.Monad.IO.Class (liftIO)
 import StrongPath ((</>))
-import Wasp.Cli.Command (Command, CommandError (..), CommandRequirement (DbConnection))
-import Wasp.Cli.Command.Common (commandRequires, findWaspProjectRootDirFromCwd)
+import Wasp.Cli.Command (Command, CommandError (..))
+import Wasp.Cli.Command.Common (CommandRequirement (DbConnection), commandRequires, findWaspProjectRootDirFromCwd)
 import Wasp.Cli.Command.Compile (compile, printWarningsAndErrorsIfAny)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Watch (watch)
